Line rendering bug fix

+++ src/main.c
@@ -39,7 +39,7 @@ void drawPixel(int x, int y, int r, int g, int b){
     glEnd();
 }
 
-void drawLine(int x1, int y1, int x2, int y2, int r, int g, int b){
+void drawLine(float x1, float y1, float x2, float y2, int r, int g, int b){
     float x = x2 - x1;
     float y = y2 - y1;
     float max = fabs(x);
